Use a non-abstract accessible role
For our viewports, we use the Widget accessible role, which is abstract.
That means that it is an overarching type representing multiple roles.
It's not very helpful to a person with vision impairment to know that
it's a general, generic widget, and because of that, GTK+ 4.12 warns
about this:

    (NeovimGtk:584029): Gtk-CRITICAL **: 21:13:28.873: gtk_widget_class_set_accessible_role: assertion '!gtk_accessible_role_is_abstract (accessible_role)' failed

To help us create a more accessible user interface and to avoid annoying
users by printing developer-specific errors to their terminal, let's
pick a more specific, non-abstract role: TextBox.